fix (filename): filename should contains the whole route from root
Right now if I have a following directory structure the with following configuration:

```
./
  ./data
     ./foo
         ./bar.json
```

```
var client = pkgcloud.storage.createClient({
  provider: 'filesystem',
  root: './'
});

client.getFiles('data', function(err, files) {
  files[0].name === 'bar.json';
});
```

As you can see the filename will only contains the basename of the file, which is wrong (or at least not compatible with other services, like S3, where the whole route from container to the file will be incuded in the filename. 

So it should be the relative position from container, `foo/bar.json`.